scanPDF_WView/requirements.txt
absl-py==0.13.0
anyio==3.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
awscli==1.19.98
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
bnbphoneticparser==0.1.5
boto3==1.17.98
botocore==1.20.98
cachetools==4.2.2
camelot-py==0.9.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
click-plugins==1.1.1
colorama==0.4.3
cryptography==3.4.7
cycler==0.10.0
darknet==0.3
darknet-py==0.2.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
docker==5.0.0
docutils==0.15.2
easydict==1.9
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.6.0
gast==0.4.0
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
gspread==3.7.0
h5py==3.1.0
httplib2==0.19.1
idna==2.10
imageio==2.9.0
imutils==0.5.4
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==8.7
keyrings.alt==4.0.2
kiwisolver==1.3.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mysql-connector==2.2.9
mysql-connector-python==8.0.25
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numpy==1.19.5
oauth2client==1.5.2
oauthlib==3.1.1
opencv-python==4.5.2.54
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pdf2image==1.15.1
pdfminer.six==20201018
peppercorn==0.6
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.18
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==1.1.2
PyYAML==5.4.1
pyzmq==22.1.0
qrcode==6.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tifffile==2021.6.14
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
voluptuous==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
wrapt==1.12.1
yolo==0.3.1
